First working iter, using gtsam gauss graph!

+++ include/kimera-vio/common/vio_types.h
@@ -47,7 +47,7 @@ enum KeypointStatus {
 using FrameId = std::uint64_t;  // Frame id is used as the index of gtsam symbol
                                 // (not as a gtsam key).
 using PlaneId = std::uint64_t;
-using LandmarkId = long int;  // -1 for invalid landmarks. // int would be too
+using LandmarkId = size_t;  // -1 for invalid landmarks. // int would be too
                               // small if it is 16 bits!
 using LandmarkIds = std::vector<LandmarkId>;
 using LandmarkCV = cv::Point3d;
